requirements.txt
numpy==1.23.2
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
Pillow
protobuf==3.20.1
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
streamlit
streamlit-drawable-canvas
supervision
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
wget==3.2
